Make the DAG processor standalone and add triggerer
We want to start using a stand-alone DAG processor for a better
separation. This commit does that. Additionally, I also added the
trigger to the scheduler container.

To achieve these two, I changed the Subprocess class such that the
execution loop (where we monitor the process and captures its logs)
manually, meaning that it is possible to create multiple parallel
subprocesses.