Support essential vs. non-essential config/environ
Some of the enviornment variables or Airlfow configuration we define are
essential and we don't want to allow the user to modify them, or others
are overridable. This commit support support having both of these, and
implement the right precedence to allow the user to override the
optional ones, but not the essential.

Additionally, I added a couple of environment variables that are
required by MWAA.