Config changes necessary for MWAA Metrics integration
I introduced a new configuration specific to MWAA for enabling
integration with CloudWatch Metrics. When the config is enabled, the
necessary Airflow configuration are set to publish StatsD metrics to
MWAA sidecar.

Additionally, when this flag is enabled, the user cannot override the
metrics configuration of Airflow, since we use those configuration to
monitor the health of the environment and, if necessary, restart the
Airflow processes or containers. Thus, to ensure the user is still able
to control the metrics they receive in their accounts, we write the
original values of the Airflow configuration the customer is trying to
override to files on a volume that is shared with the sidecar, and then
let the sidecar decide what to publish to the customer.