Make auto scaling metrics report zero
Currently, our code doesn't publish a metric in case there is nothing to
report, instead of reporting 0. This commit changes it such that we
always report a value. This helps us confirm that the auto scaling code
is working, just there is nothing to do, as opposed to seeing no metric,
which makes it difficult to know when there is an issue with our logic.