Reduce CONSECUTIVE_IDLENESS_CHECK_THRESHOLD to 2
Reducing `CONSECUTIVE_IDLENESS_CHECK_THRESHOLD` from 15 to 2 to avoid
lengthy wait before shutting down an idle worker. This setting, when
applied to the internal code, resulted in ~ 15 seconds of delay, but the
new open source image are designed differently and our checks are
executed once 1 minute, so 15 consecutive checks means 15 minutes, which
is considerable time. So, reducing this to 2 to keep the idleness wait
to roughly 2 minutes.