Improvement and some fixes for Subprocess class + Disable "resetdb"
Multiple improvements and bug fixes on the Subprocess class and its
related code:

- Improved and generic mechanism for handling SIGTERMs that ensure the
  subprocess and its resources are properly cleaned.
- Refactoring and improved logging.

Also, I disabled the `resetdb` command for now, as it is pretty risky to
have such a destructive command adjacent to other commands used in
production; a simple code mistake can result in wiping out production
databases. Instead, testing commands like this should be in a completely
isolated boundary, with protection mechanism to ensure they are not
accidentally executed in production.