Bill of Materials for Docker Image distribution (#116)
*Issue #, if available:* N/A

*Description of changes:*

Created scripts to automate the building of Bill of Materials for Docker
Image distribution. This PR achieves two things:

- It adds a bootstrap step to generate the Bill of Materials for both,
OS and Python packages, and dump the content under the
`/BillsOfMaterials` folder in the Docker Image for visibility.
- It syncs the generated files locally in the repo code so it gets
automatically added in source control. This way, future updates to the
used packages will automatically result in updating the files in the
repo, thus making it easy to review as well.

---

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.