Fix issue: Sentry causes failures in 2.9.2 (#122)
*Issue #, if available:* #121

*Description of changes:*

Fix issue #121. The issue is caused because we are trying to initialize
the DB before installing the requirements. This causes this line in
Airflow code base to fail because `sentry_sdk` isn't installed yet:

https://github.com/apache/airflow/blob/2.9.2/airflow/sentry.py#L58

To fix this issue, I moved the installation of PIP requirements before
the DB initialization.

This is unrelated to this issue, but when creating a process to install
the requirements, there is no need to pass the environment variables we
craft for Airflow processes, as what we simply just running a pip
install command.

---

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.