Fix issue 123: Repeated reserialization of DAGs
Issue #124 happens because we are always calling `airflow db migrate` to
make sure the database is migrated. The assumption was that the
operation is a no-op if the DB is already initialized. However, this is
not the case, and a reserialization of DAGs is happening every time this
command is called. To avoid this, I am changing the code for executing
migration to check if migration is needed.