Add hybrid scheduler/worker container (#119)
Adds the command 'hybrid' to the entrypoint script to allow creation of
a container that runs both celery worker and scheduler sub-processes.

*Issue #, if available:* N/A

*Description of changes:*

Introduces a new supported command 'hybrid' to run celery worker and
scheduler sub-processes in the same docker container. Health-check
conditions for the 'hybrid' container are limited to the 'worker '
sub-process because the Sidecar does not support multiple health-check
simultaneous connections at this time. This 'hybrid' container uses the
`CeleryExecutor`, but we plan to add a separate container command to
support a `LocalExecutor` setup in the future.


---

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Kyle Thatcher <kytha@amazon.com>