Fix issue 123: Repeated reserialization of DAGs (#125)
*Issue #, if available:* #124

## Description of changes

Issue #124 happens because we are always calling `airflow db migrate` to
make sure the database is migrated. The assumption was that the
operation is a no-op if the DB is already initialized. However, this is
not the case, and a reserialization of DAGs is happening every time this
command is called. To avoid this, I am changing the code for executing
migration to check if migration is needed.

## Testing

I did multiple tests using the Docker Compose setup. More testing by
Amazon MWAA developers before this PR is merged. Suggested testing:

1. Create an internal 2.9.2 environment and make sure it is functioning
normally. Also check ECS logs to ensure there are no issues.
2. Create a pre-2.9.2 environment and do an in-place upgrade to 2.9.2
and make sure the process is flawless.
3. Run a 2.9.2 environment with auto scaling enabled (ensuring that
workers are being created and terminated continuously) and make sure
migrations are not causing any issues.

---

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.