Adding minor details to the build and run step within the README inst… (#128)
*Description of changes:*

Adding a couple of additional details to the build and run step. The
run.sh file require a few changes before it is ready to be executed and
these steps are not present in the README file.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.