Refactor logging to be async (#134)
*Description of changes:*

Performance testing MWAA airflow images revealed flaws with how logging
is plumbed to CloudWatch.

Currently, when airflow publishes log statements it is sent to an os
pipe to be polled and read by the entrypoint process and then forwarded
to cloudwatch. When this pipe is full, airflow must halt until logs are
read before continuing execution. This means airflow execution is
upper-bounded by how fast the entrypoint can read logs.

The entrypoint has many other priorities in this critical loop
(heartbeat healthcheck, db healthcheck, polling other subprocess logs,
etc) and it has be measured that airflow scheduler is significantly
throttled by this mechanism when log volume is large.

This PR aims to fix this by separating log polling into its own thread.
This allows for faster log reading that is decoupled from the execution
of the entrypoint process loop. Below is "before" and "after" profiler
flamegraph results of a mwaa scheduler over a 1-minute period of
execution under heavy scheduling and log load.

### Before
![Screenshot 2024-08-27 at 1 43
08 AM](https://github.com/user-attachments/assets/d2ecf5f9-7260-4de3-9654-ac0a48cee457)


### After
![Screenshot 2024-08-27 at 1 40
26 AM](https://github.com/user-attachments/assets/dddfc792-54c3-47a9-9f6d-20c5927b3088)

In the before, we see that waiting for the log message to be read by the
entry point is all the scheduler was able to do in the entire 1-minute
interval. However with async logging, scheduler is able to complete
multiple loops.

---------

Co-authored-by: Kyle Thatcher <kytha@amazon.com>