Applying 12 hour limit from time of processing termination signal
When a termination signal is received by a worker, then it is provided TERMINATION_TIME_LIMIT amount of time to graceful shutdown by finishing up the current Airflow tasks. But if termination signals are received late due to an issue, then we need to allow the TERMINATION_TIME_LIMIT to start from the point in time of processing the signal and not the time when the signal was sent from MWAA. This change keeps track of the last time the termination signal was processed which is then later used to check if the 12 hour mark has been breached or not.

This change also introduces minor failsafe related to signal file processing and reduces sleep time between each time the Airflow task consumption is paused/resumed.