feat: support LocalExecutor (#138)
1. Adds support for LocalExecutor when `MWAA__SQS__CREATE_QUEUE` is set
to false. If we're not creating an SQS queue/using Celery, then we don't
want to attempt to create the CeleryExecutor. This is useful for
environments where a discrete worker is not necessary/too expensive.
2. Adds a config variable `MWAA__CORE__DISABLE_CORS` (default false),
which, when set to true, sets all CORS-related headers to accept all
connections. This is required to make MWAA work in embedded environments
(i.e. within a portal or webpage)