Porting graceful updates image changes to 2.10.1
The task monitor needs to process the following 4 MWAA signals for the graceful update project:

1. Termination signal: Graceful termination of the workers when the environment is going through a graceful update
2. Resume signal: Reverting the state of graceful termination and resume work when the environment is going through a rollback after attempting a graceful update
3. Kill signal: Shutting down the worker without waiting for the current Airflow tasks to finish when the environment is going through a forced update
4. Activation signal: Starting consumption of work from the queue after termination protection has enabled on the corresponding Fargate task

The processing is gated behind certain environment variables which are either absent or marked as false for an environment which does not have graceful updates enabled.

The CR brings the changes that have been merged for 2.9.2 (https://github.com/aws/amazon-mwaa-docker-images/pull/137 and https://github.com/aws/amazon-mwaa-docker-images/pull/150) to the newer version.