Add test-requirements,test-startup-script, and resetdb script commands (#213)
*Issue #, if available:*

*Description of changes:*
Added 3 new commands for testing similar to those from local runner.
They can be ran like`./run.sh test-requirements`. Full instructions are
in the readme. Purpose of them is to test dependencies without starting
airflow.

Note: for resetdb, you have to manually uncomment a section in your
versions entrypoint as we do not want this section enabled in production
as having it execute on a customer's database would reset it

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Michael Fu <miklfu@amazon.com>