Reserialize dags on db migration (#218)
*Issue #, if available:*

*Description of changes:*
In previous airflow versions, we had reserialize dags as true. Right now
it does not so setting this config to True. Tested by having the run
script go to this path and successfully execute the command with this
new config.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Michael Fu <miklfu@amazon.com>