Create log groups in customer account if not exists. #129 (#219)
*Issue #, if available:#129*

*Description of changes:*
Updated the `run.sh` script to check if the log group exists in the
customer account or not. If it does not exists, create the required log
group for all airflow components.

*Testing*
1. Checked if the log groups were created in the account or not when
running the script.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

Co-authored-by: sudiptoghosh01 <hsudipto@amazon.com>