Update generated docker artifacts (#244)
*Description of changes:*

Updating BOM artifacts and generated dockerfile. This was causing an
issue where the dockerfile template was modified in [this
PR](https://github.com/aws/amazon-mwaa-docker-images/pull/233), but the
generated artifact was not included, so downstream build systems were
not getting latest dockerfile.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

Co-authored-by: Kyle Thatcher <kytha@amazon.com>