Add migrate-db command and support airflow downgrades (#247)
*Issue #, if available:*

*Description of changes:*
Added a new command `migrate-db` for a special container that will run
the same airflow db init/upgrade logic but will also check for an
environment variable `MWAA__CORE__TARGET_VERSION` to downgrade. Existing
inplace upgrade will not be changed for now.

The migrate_with_downgrade.py file is a copy or migrate.py but with an
additional downgrade check

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Michael Fu <miklfu@amazon.com>