Fix logging issue (#271)
*Issue*:
When the log level is set in AWS Console for Worker and Webserver as
**ERROR or WARN**, some **INFO** logs still appear in CloudWatch.

*Description of changes*:
The problem seems to be around Airflow where the log level set in MWAA
is not adaopted by Airflow correctly or Airflow requires additional
configuration to handle this.
In order to fix this issue on MWAA side, we are verifying the LOG LEVEL
set in console and the LOG LEVEL of the log message.

***Note:*** Quality checks are failing due to missing test cases for
version 3.0.1