Add new task logging handler for Airflow 3, plus minor task API related setup changes (#287)
Revamped the MWAA custom task logger for the Airflow 3 Structlog era. 

Since task logging will be carried out with Structlog in Airflow 3, our
custom logger needs to have the core upload logic implemented in the
processor property which is only loaded if remote logging is enabled
before task execution starts. (in supervisor.py)

In the logging config setup, we also need to initialize the remote
logging class and pass it to the REMOTE_TASK_LOG variable which will be
loaded by Airflow.

Minor changes to the environment variable setup to enable task api.

Unit test will be added in a future PR.

---------

Co-authored-by: Suyuan Yu <suyuay@amazon.com>