Remove Deprecated Config & SimpleAuthManager Username Fix (#312)
*Description of changes:*

- **Fix default username for SimpleAuthManager:** There was a bug where
both admin and viewer roles had the same username. This is of the format
`username:role`. Ref:
https://airflow.apache.org/docs/apache-airflow/stable/configurations-ref.html#simple-auth-manager-users

- **Remove deprecated AIRFLOW__WEBSERVER__SECRET_KEY config from Airflow
3.0.3:** `AIRFLOW__WEBSERVER__SECRET_KEY` no longer exists in Airflow 3.
`AIRFLOW__API__SECRET_KEY` is the equivalent which is added. Ref:
https://airflow.apache.org/docs/apache-airflow/stable/configurations-ref.html#secret-key


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.