Removing FAB auth-disabling irrelevant in AF3 and adding configs to protected env var list (#316)
*Issue #, if available:*

*Description of changes:*
- In Airflow 3, for production we set `MWAA__CORE__AUTH_TYPE` to
"mwaa-iam" with `FAB` as the Auth Manager , and during local testing via
`run.sh`, we set it to "testing", and use the `Simple Auth Manager`, not
FAB. The code in `_get_essential_airflow_api_config()` is thus not
needed in Airflow 3, and also setting `AUTH_ROLE_PUBLIC` is not useful
since that is a FAB config, hence removed them, since they are used to
disable auth if we were using FAB.

- Adding `AIRFLOW__API__BASE_URL` and
`AIRFLOW__CORE__DATABASE_ACCESS_ISOLATION` to the the protected os
environment variable list.
- Refer
[link1](https://airflow.apache.org/docs/apache-airflow/2.9.1/configurations-ref.html#auth-backends)
and
[link2](https://airflow.apache.org/docs/apache-airflow-providers-fab/1.1.1/auth-manager/api-authentication.html)
for documentation on how the old code was used to disable auth.

Tested this change by deploying locally and verifying Auth works as
expected.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

Co-authored-by: Kashyap Kannan <kashkann@amazon.com>