Set reserialize_dags to False on DB upgrades and downgrades (#313)
*Issue #, if available:*

*Description of changes:*
Set reserialize_dags to False on DB upgrades and downgrades as it is
loading dags for the upgrades/downgrades. We do not install requirements
for this process now so we cannot load their dags.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Michael Fu <miklfu@amazon.com>