Disable dag serialization when performing database migration (#325)
*Issue #, if available:* n/a

*Description of changes:*

Disables dag serialization during Airflow db migrations for the
migrate-db command. The command is intended to be used without custom
plugins/requirements/startup script that could interfere with db
operations. This means that it is unable to serialize dags which require
dependencies which have not been installed. So we need to disable dag
serialization during the migration. This is ok because the dag processor
is responsible for dag serialization and will handle it after the db
migration/initialization. The flag was removed from Airflow as of
Airflow 2.11.0 because the dag processor makes it redundant:
https://github.com/apache/airflow/pull/45362.

For testing I deployed a control environment and test environment to
verify the dag serialization errors are resolved by the fix. The dags
functioned as expected with the flag disabled.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.