Remove migrate being ran on scheduler and webserver (#341)
*Issue #, if available:*

*Description of changes:*
Removing the migrate command from being ran on schedulers and webservers
because it is already performed on the migratedb container. This will
prevent database corruptions in the future as leftover
schedulers/webservers on rollbacks cause this issue.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

Co-authored-by: Michael Fu <miklfu@amazon.com>