Use Empty Config for MigrateDb container (#360)
*Issue #, if available:*

*Description of changes:*

If a customer passed in a config that required a plugin installation, it
can cause errors in the migratedb container because we do not sync
plugins for this container. Using an empty config to allow this process
to be more isolated so the customer cannot break it. This does not
affect other containers. Tested by using this image for a MWAA
environment and verified that the config no longer breaks it.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

Co-authored-by: Michael Fu <miklfu@amazon.com>