Update error message for failed RDS credentials (#364)
*Issue #, if available:*
https://github.com/aws/amazon-mwaa-docker-images/issues/363
*Description of changes:*
Updating the error log to a warning log. This RDS credential code
doesn't work locally so we skip if we encounter an error as it is not
necessary for a local run.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

Co-authored-by: Michael Fu <miklfu@amazon.com>