Revert "Increase pool size for problematic sizes (#358)" (#371)
This reverts commit 48d052eeb72f5035233c6d1b9b0312745fe99038.

*Issue #, if available:*

*Description of changes:*
Reverting this change as it is causing unnecessary confusion as to why
pool size config is getting reset on update. MWAA is now setting the
pool size to 10000 on create which means the service itself is not
adding an unnecessary bottle neck on the number of concurrent tasks
which can run in an environment. Customer can feel free to change the
pool size as desired.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.